+++ zfit/constraint.py
@@ -3,10 +3,10 @@
 import tensorflow as tf
 
 from .core.constraint import (GaussianConstraint, PoissonConstraint,
-                              SimpleConstraint)
+                              SimpleConstraint, LogNormalConstraint)
 from .util import ztyping
 
-__all__ = ["nll_gaussian", "SimpleConstraint", "GaussianConstraint", "PoissonConstraint"]
+__all__ = ["nll_gaussian", "SimpleConstraint", "GaussianConstraint", "PoissonConstraint", "LogNormalConstraint"]
 
 
 def nll_gaussian(params: ztyping.ParamTypeInput, observation: ztyping.NumericalScalarType,